Tz/ Try to fix import errors when running files for testing
