+++ .vscode/settings.json
@@ -1,4 +1,4 @@
 {
     "python.linting.flake8Enabled": true,
-    "python.linting.enabled": true
+    "python.linting.enabled": false
 }
