+++ App/server/database.py
@@ -1,10 +1,9 @@
-import motor.motor_asyncio
 import os
+
+import motor.motor_asyncio
 from dotenv import load_dotenv
 
 load_dotenv()
 
 client = motor.motor_asyncio.AsyncIOMotorClient(os.environ["MONGODB_URL"])
 db = client.hydro
-
-fefe = 3
