+++ App/server/models/garden.py
@@ -1,10 +1,18 @@
+import os
+import sys
 from datetime import datetime
 from typing import Optional
 
 from bson import ObjectId
+
+parent = os.path.abspath(".")
+sys.path.append(parent)
+
+# print(sys.path)
 from dotenv import load_dotenv
 from pydantic import BaseModel, Field, root_validator
-from .id import PyObjectId
+
+from App.server.models.id import PyObjectId
 
 load_dotenv()
