+++ App/server/models/scheduled_actuator.py
@@ -1,11 +1,16 @@
+import os
+import sys
 from datetime import datetime
 from typing import Union
+
+parent = os.path.abspath(".")
+sys.path.append(parent)
 
 from bson import ObjectId
 from dotenv import load_dotenv
 from pydantic import BaseModel, Field, root_validator
 
-from .id import PyObjectId
+from App.server.models.id import PyObjectId
 
 load_dotenv()
