+++ App/server/models/schedules.py
@@ -1,8 +1,13 @@
+import os
+import sys
 from typing import List
 
 from pydantic import BaseModel, Field
 
-from .id import PyObjectId
+parent = os.path.abspath(".")
+sys.path.append(parent)
+
+from App.server.models.id import PyObjectId
 
 
 class SS(BaseModel):
