+++ App/server/models/sensor.py
@@ -1,19 +1,25 @@
+import os
+import sys
 from datetime import datetime
-from typing import Union
+from typing import Optional, Union
 
 from bson import ObjectId
 from dotenv import load_dotenv
 from pydantic import BaseModel, Field, root_validator
 
-from .id import PyObjectId
+parent = os.path.abspath(".")
+sys.path.append(parent)
+
+from App.server.models.id import PyObjectId
 
 load_dotenv()
 
 
-class ReactiveActuatorModel(BaseModel):
+class SensorModel(BaseModel):
     id: PyObjectId = Field(default_factory=PyObjectId, alias="_id")
     name: str = Field(...)
-    sensor_id: Union[PyObjectId, None] = Field(default=None, alias="sensor_id")
+    garden_id: Union[PyObjectId, None] = Field(default=None, alias="garden_id")
+
     created_at: datetime = datetime.now()
     updated_at: datetime = datetime.now()
 
@@ -23,13 +29,20 @@ class ReactiveActuatorModel(BaseModel):
         validate_assignment = True
         json_encoders = {ObjectId: str}
         schema_extra = {
-            "example": {
-                "name": "Nutrient pump",
-                "sensor_id": "6359d55bff77b777dd5c92e8",
-            }
+            "example": {"name": "Temperature", "garden_id": "6359d55bff77b777dd5c92e8"}
         }
 
         @root_validator
         def number_validator(cls, values):
             values["updated_at"] = datetime.now()
             return values
+
+
+class UpdateSensorModel(BaseModel):
+    name: Optional[str]
+    garden_id: Optional[PyObjectId]
+
+    class Config:
+        arbitrary_types_allowed = True
+        json_encoders = {ObjectId: str}
+        schema_extra = {"example": {"name": "Humidity"}}
